CHANGES: 1) added single-precision floating point version og natural logarithm function  - log_functions_.asm 2) added log_sse_packed float version of function  - log_functions.h 3) added unit test for point 1) and 2)
TODO:
- add double-precision floating point verison of natural logarithm
function

+++ unit_tests/exp_functions_t.h
@@ -84,9 +84,9 @@ void testBasicExpSSEFloat() {
 	int const n = 16;
 	std::size_t const align = 16;
 
-	double* x = aligned_alloc<double>(n, align);
-	double* res1 = aligned_alloc<double>(n, align);
-	double* res2 = aligned_alloc<double>(n, align);
+	float* x = aligned_alloc<float>(n, align);
+	float* res1 = aligned_alloc<float>(n, align);
+	float* res2 = aligned_alloc<float>(n, align);
 
 	// test some basic known values:
 
@@ -201,9 +201,9 @@ void testBasicFastExpSSEFloat() {
 	int const n = 16;
 	std::size_t const align = 16;
 
-	double* x = aligned_alloc<double>(n, align);
-	double* res1 = aligned_alloc<double>(n, align);
-	double* res2 = aligned_alloc<double>(n, align);
+	float* x = aligned_alloc<float>(n, align);
+	float* res1 = aligned_alloc<float>(n, align);
+	float* res2 = aligned_alloc<float>(n, align);
 
 	// test some basic known values:
